+++ iceoryx_examples/waitset/ice_waitset_trigger.cpp
@@ -16,6 +16,7 @@
 #include "iceoryx_posh/runtime/posh_runtime.hpp"
 #include "iceoryx_utils/cxx/optional.hpp"
 
+#include <bits/stdint-uintn.h>
 #include <iostream>
 #include <thread>
 
@@ -47,7 +48,7 @@ class MyTriggerClass
     MyTriggerClass& operator=(MyTriggerClass&&) = delete;
 
     // When you call this method you will trigger the ACTIVATE event
-    void activate(const int activationCode) noexcept
+    void activate(const uint64_t activationCode) noexcept
     {
         m_activationCode = activationCode;
         m_isActivated = true;
@@ -224,8 +225,8 @@ int main()
 
     // start a thread which will trigger a event every second
     std::thread triggerThread([&] {
-        int activationCode = 1;
-        for (auto i = 0; i < 10; ++i)
+        uint64_t activationCode = 1U;
+        for (auto i = 0U; i < 10; ++i)
         {
             std::this_thread::sleep_for(std::chrono::seconds(1));
             triggerClass->activate(activationCode++);
